+++ libraries/common/src/main/java/androidx/media3/common/util/MediaFormatUtil.java
@@ -252,7 +252,7 @@ public final class MediaFormatUtil {
    */
   @SuppressWarnings("InlinedApi")
   public static void maybeSetColorInfo(MediaFormat format, @Nullable ColorInfo colorInfo) {
-    if (colorInfo != null) {
+    if (!ColorInfo.isEquivalentToAssumedSdrDefault(colorInfo)) {
       maybeSetInteger(format, MediaFormat.KEY_COLOR_TRANSFER, colorInfo.colorTransfer);
       maybeSetInteger(format, MediaFormat.KEY_COLOR_STANDARD, colorInfo.colorSpace);
       maybeSetInteger(format, MediaFormat.KEY_COLOR_RANGE, colorInfo.colorRange);