+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/mediacodec/MediaCodecInfo.java
@@ -41,6 +41,7 @@ import android.util.Pair;
 import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
 import androidx.annotation.VisibleForTesting;
+import androidx.media3.common.ColorInfo;
 import androidx.media3.common.Format;
 import androidx.media3.common.MimeTypes;
 import androidx.media3.common.util.Assertions;
@@ -422,7 +423,10 @@ public final class MediaCodecInfo {
           && (oldFormat.width != newFormat.width || oldFormat.height != newFormat.height)) {
         discardReasons |= DISCARD_REASON_VIDEO_RESOLUTION_CHANGED;
       }
-      if (!Util.areEqual(oldFormat.colorInfo, newFormat.colorInfo)) {
+      if ((!ColorInfo.isEquivalentToAssumedSdrDefault(oldFormat.colorInfo)
+              || !ColorInfo.isEquivalentToAssumedSdrDefault(newFormat.colorInfo))
+          && !Util.areEqual(oldFormat.colorInfo, newFormat.colorInfo)) {
+        // Don't perform detailed checks if both ColorInfos fall within the default SDR assumption.
         discardReasons |= DISCARD_REASON_VIDEO_COLOR_INFO_CHANGED;
       }
       if (needsAdaptationReconfigureWorkaround(name)