+++ RELEASENOTES.md
@@ -50,6 +50,9 @@
     *   Add workaround that ensures the first frame is always rendered while
         tunneling even if the device does not do this automatically as required
         by the API ([#1169](https://github.com/androidx/media/issues/1169)).
+    *   Fix issue where HDR color info handling causes codec mishavior and
+        prevents adaptive format switches for SDR video tracks
+        ([#1158](https://github.com/androidx/media/issues/1158)).
 *   Text:
     *   WebVTT: Prevent directly consecutive cues from creating spurious
         additional `CuesWithTiming` instances from `WebvttParser.parse`